add some documents and move joystick over to BLE component
