+++ Readme.md
@@ -1,2 +1,2 @@
 Generate Docs For the BLE Component
-idf.py BLE_DOCS
+doxygen doxyfile
