+++ components/BLE/CMakeLists.txt
@@ -1,4 +1,4 @@
-idf_component_register(SRCS "ble.cpp"  "device.cpp" "service.cpp" "characteristic.cpp"
+idf_component_register(SRC_DIRS "./"
                        INCLUDE_DIRS "./include"
                        REQUIRES bt driver esp_common nvs_flash)
