+++ components/Car/include/car.hpp
@@ -8,12 +8,28 @@
 class Car
 {
 public:
+    /// @brief Construct a Car that has 2 speeds cruise(set through functions) and Turbo(max speed)
+    /// @param remote Remote that controls the car
+    /// @param leftMotor - left motor
+    /// @param rightMotor - right motor
     Car(std::shared_ptr<Mocute052> remote, Motor *leftMotor, Motor *rightMotor);
 
+    /**
+     * @brief Get the Cruise Speed of the car
+     *
+     * @return float - Current cruise speed of the car
+     */
     float getCruiseSpeed() { return mMotorMixer->getCruiseSpeed(); };
+    /**
+     * @brief Set the max speed of the car when turbo is not enabled
+     *
+     * @param aCruiseSpeed - a speed that will be max without turbo mode enabled
+     */
     void setCruiseSpeed(float aCruiseSpeed) { return mMotorMixer->setCruiseSpeed(aCruiseSpeed); };
 
+    ///@brief Allow to car to go full speed
     void enableTurbo() { mMotorMixer->enableTurbo(); };
+    /// @brief Force car to use cruising speed as the fastest speed.
     void disableTurbo() { mMotorMixer->disableTurbo(); };
 
 private:
@@ -25,7 +41,9 @@ private:
      */
     void setMotorSpeed(float aLeftMotorSpeed, float aRightMotorSpeed);
 
+    /// @brief Refences to the right motor
     std::unique_ptr<Motor> mRightMotor;
+    /// @brief Reference to Left motor
     std::unique_ptr<Motor> mLeftMotor;
 
     /**
